requirements.txt
bcrypt==3.1.4
cffi==1.11.4
Django==2.0.2
django-bootstrap3==9.1.0
django-environ==0.4.4
mysqlclient==1.3.12
pycparser==2.18
pytz==2018.3
six==1.11.0

django-enumfields==0.9.0
django-embed-video==1.1.0
